Conflicts:
	src/XMakeBuildEngine/UnitTests/BackEnd/BuildManager_Tests.cs

+++ src/Utilities/ToolTask.cs
@@ -60,6 +60,8 @@ namespace Microsoft.Build.Utilities
     /// </remarks>
     public abstract class ToolTask : Task, ICancelableTask
     {
+        private static bool s_preserveTempFiles = String.Equals(Environment.GetEnvironmentVariable("MSBUILDPRESERVETOOLTEMPFILES"), "1", StringComparison.Ordinal);
+
         #region Constructors
 
         /// <summary>
@@ -944,6 +946,12 @@ namespace Microsoft.Build.Utilities
         /// <param name="filename">File to delete</param>
         protected void DeleteTempFile(string fileName)
         {
+            if (s_preserveTempFiles)
+            {
+                Log.LogMessageFromText($"Preserving temporary file '{fileName}'", MessageImportance.Low);
+                return;
+            }
+
             try
             {
                 File.Delete(fileName);
@@ -1688,7 +1696,7 @@ namespace Microsoft.Build.Utilities
                 // Clean up after ourselves.
                 if (_temporaryBatchFile != null && File.Exists(_temporaryBatchFile))
                 {
-                    File.Delete(_temporaryBatchFile);
+                    DeleteTempFile(_temporaryBatchFile);
                 }
             }
         } // Execute()
